Add a type casting hierarchy for merging input mappings together.
Additionally, the scaled float should be treated as a double value in
Hadoop.

Fixes #1192, fixes #1074

+++ spark/sql-20/src/main/scala/org/elasticsearch/spark/sql/SchemaUtils.scala
@@ -160,7 +160,7 @@ private[sql] object SchemaUtils {
       case FLOAT        => FloatType
       case DOUBLE       => DoubleType
       case HALF_FLOAT   => FloatType
-      case SCALED_FLOAT => FloatType
+      case SCALED_FLOAT => DoubleType
       // String type
       case STRING       => StringType
       case TEXT         => StringType
